pyproject.toml
[project]
name = "phidata-youtube-summarizer"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "phidata>=2.4.37",
    "groq>=0.11.0",
    "streamlit>=1.38.0",
    "youtube-transcript-api>=0.6.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/phidata_youtube_summarizer"]
